[web] Rename sign-pages to auth
