+++ src/pages/auth/index.ts
